rename: complete renaming RenameInputField to RenameWidget
