+++ src/vs/editor/contrib/rename/browser/renameWidget.css
